// server/src/types/entities.ts
export interface DashboardMetricsDTO {
  arrivalsToday: number;
  departuresToday: number;
  currentOccupancyPct: number;
  inHouseGuests: number;
  totalRooms: number;
  occupiedRooms: number;
  availabilityMatrix: {
    dates: string[];
    categories: Array<{
      id: string;
      code: string;
      name: string;
      total: number;
      rates: number[];
      counts: number[];
    }>;
  };
}